+++ src/main/java/Parser.java
@@ -1,3 +1,4 @@
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface Parser {
@@ -5,7 +6,7 @@ public interface Parser {
 
 
     public void parse(String expression);
-    public List<Double> getOperands();
+    public List<BigDecimal> getOperands();
     public char getOperator();
 
 }
